Refactor agent runtime into modular architecture
