+++ tenvy-client/internal/agent/seed.go
@@ -1,0 +1,10 @@
+package agent
+
+import (
+	"math/rand"
+	"time"
+)
+
+func init() {
+	rand.Seed(time.Now().UnixNano())
+}
