+++ tenvy-client/internal/agent/util.go
@@ -1,0 +1,10 @@
+package agent
+
+import "time"
+
+func minDuration(a, b time.Duration) time.Duration {
+	if a < b {
+		return a
+	}
+	return b
+}
